Main navigation link padding changes, image upload margin changes, and image replacement added for Ideastrike branding.
Nudged the main navigation links down a bit by increasing the padding.
This lines up the bottom of the heading and main nav link text.

Reduced the margin on the image upload form control so it's left edge
lines up with the other form controls.

Changed the padding and the height of the .topbar .brand CSS to use LIR
for image replacement. Added <h1> tag to nested inside the .brand <a>
and added the word Ideastrike (<h1> can be nested inside <a> here as
the doctype is HTML5).